requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
hvac==2.1.0
idna==3.6
multi-key-dict==2.0.3
pbr==6.0.0
python-jenkins==1.8.2
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
urllib3==2.2.1
